+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * use `qs2` package instead of `qs` (deprecated) for improved object serialization (#291; #316);
 * fix warning capture in `spades()` debug logging;
+* fix use of `debug` as `verbose` when caching (#322);
 * documentation improvements;
 
 # SpaDES.core 2.1.8